extracts upvote card to container
